refactor(agent): extract publish tool handler
